requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
werkzeug>=3.0.0
Pillow>=10.0.0
gunicorn>=21.2.0

